# Conflicts:
#	CHANGELOG.md
#	platforms/joomla/install.php
#	platforms/joomla/updates/pkg_gantry5.xml
#	platforms/joomla/updates/tpl_g5_helium.xml
#	platforms/joomla/updates/tpl_g5_hydrogen.xml

+++ src/classes/Gantry/Component/Filesystem/Folder.php
@@ -340,7 +340,7 @@ abstract class Folder
     protected static function doDelete($folder, $include_target = true)
     {
         // Special case for symbolic links.
-        if (is_link($folder)) {
+        if ($include_target && is_link($folder)) {
             return @unlink($folder);
         }
